+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/BlinkSpell.java
@@ -1,10 +1,10 @@
 package com.elmakers.mine.bukkit.plugins.spells.builtin;
 
-import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.World;
+import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
 
 import com.elmakers.mine.bukkit.plugins.spells.Spell;
 import com.elmakers.mine.bukkit.plugins.spells.utilities.PluginProperties;