+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/ConstructSpell.java
@@ -3,10 +3,8 @@ package com.elmakers.mine.bukkit.plugins.spells.builtin;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
-import org.bukkit.entity.Entity;
 import org.bukkit.inventory.ItemStack;
 
 import com.elmakers.mine.bukkit.persistence.dao.BlockList;
@@ -22,7 +20,6 @@ public class ConstructSpell extends Spell
 	private int				defaultRadius			= 2;
 	private int				maxRadius				= 32;
 	private int				defaultSearchDistance	= 32;
-    private String          targetType              = "block";
 	
 	public ConstructSpell()
 	{
@@ -57,26 +54,15 @@ public class ConstructSpell extends Spell
 	{
 		setMaxRange(defaultSearchDistance, true);
 		targetThrough(Material.GLASS);
-		Block target = null;
-		Entity targetEntity = getTargetEntity();
-		if (targetEntity != null)
-		{
-		    Location targetLocation = targetEntity.getLocation();
-		    target = targetLocation.getBlock();
-		    targetType = targetEntity.getClass().getName().toLowerCase();
-		}
-		else
-		{
-		    targetType = "block";
-		    target = getTargetBlock();
-		    if (target == null)
-		    {
-	            initializeTargeting(player);
-	            noTargetThrough(Material.GLASS);
-	            target = getTargetBlock();
-	        }
-		}
+		Block target = getTarget().getBlock();
 		
+	    if (target == null)
+	    {
+            initializeTargeting(player);
+            noTargetThrough(Material.GLASS);
+            target = getTarget().getBlock();
+        }
+	
         if (target == null)
         {
             castMessage(player, "No target");